session

+++ service_diagnostic_session.go
@@ -2,10 +2,10 @@ package uds
 
 //import "fmt"
 
-func service_diagnostic_session_make_request(session byte) *Request {
+func service_diagnostic_session_make_request(session int) *Request {
 	r := NewRequest(0x10, "diagnostic_session")
 	r.use_subfunction = true
-	r.subfunction = session
+	r.subfunction = byte(session)
 	return r
 }
